ScenarioBench-CA_Coverage_Kit/nlq/requirements.txt
pydantic>=2.7
jsonschema>=4.21
PyYAML>=6.0
